a major improvement of the descriptions, thanks to `phpstan`

+++ src/Controller/SystemsController.php
@@ -116,7 +116,7 @@ class SystemsController extends AppController
     {
         //Checks if the sitemap exist and is not expired
         if (is_readable(SITEMAP)) {
-            $time = Time::createFromTimestamp(filemtime(SITEMAP));
+            $time = Time::createFromTimestamp((int)filemtime(SITEMAP));
 
             if (!$time->modify(getConfigOrFail('main.sitemap_expiration'))->isPast()) {
                 $sitemap = file_get_contents(SITEMAP);
